Add files via upload
Run as Administrator:
Right-click on the .bat file.
From the context menu, select Run as administrator. This is necessary for the script to execute with the required permissions.

User Account Control (UAC):
If prompted by User Account Control, click Yes to allow the batch file to make changes to your device.